+++ src/Tag/Base/ContentTag.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Html\Tag\Base;
+
+abstract class ContentTag extends NormalTag implements ContentTagInterface
+{
+    use TagContentTrait;
+}