+++ src/Tag/Base/VoidTag.php
@@ -22,6 +22,11 @@ abstract class VoidTag extends Tag
         return new static();
     }
 
+    final public function getContent(): string
+    {
+        return '';
+    }
+
     final protected function renderTag(): string
     {
         return '<' . $this->getName() . $this->renderAttributes() . '>';