+++ src/Tag/Caption.php
@@ -4,16 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Html\Tag;
 
-use Yiisoft\Html\Tag\Base\NormalTag;
-use Yiisoft\Html\Tag\Base\TagContentTrait;
+use Yiisoft\Html\Tag\Base\ContentTag;
 
 /**
  * @link https://www.w3.org/TR/html52/tabular-data.html#the-caption-element
  */
-final class Caption extends NormalTag
+final class Caption extends ContentTag
 {
-    use TagContentTrait;
-
     protected function getName(): string
     {
         return 'caption';