+++ src/Tag/I.php
@@ -4,16 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Html\Tag;
 
-use Yiisoft\Html\Tag\Base\NormalTag;
-use Yiisoft\Html\Tag\Base\TagContentTrait;
+use Yiisoft\Html\Tag\Base\ContentTag;
 
 /**
  * @link https://www.w3.org/TR/html52/textlevel-semantics.html#the-i-element
  */
-final class I extends NormalTag
+final class I extends ContentTag
 {
-    use TagContentTrait;
-
     protected function getName(): string
     {
         return 'i';