+++ src/Tag/Label.php
@@ -4,16 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Html\Tag;
 
-use Yiisoft\Html\Tag\Base\NormalTag;
-use Yiisoft\Html\Tag\Base\TagContentTrait;
+use Yiisoft\Html\Tag\Base\ContentTag;
 
 /**
  * @link https://www.w3.org/TR/html52/sec-forms.html#the-label-element
  */
-final class Label extends NormalTag
+final class Label extends ContentTag
 {
-    use TagContentTrait;
-
     public function forId(?string $id): self
     {
         $new = clone $this;