+++ src/Tag/Noscript.php
@@ -4,16 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Html\Tag;
 
-use Yiisoft\Html\Tag\Base\NormalTag;
-use Yiisoft\Html\Tag\Base\TagContentTrait;
+use Yiisoft\Html\Tag\Base\ContentTag;
 
 /**
  * @link https://html.spec.whatwg.org/multipage/scripting.html#the-noscript-element
  */
-final class Noscript extends NormalTag
+final class Noscript extends ContentTag
 {
-    use TagContentTrait;
-
     protected function getName(): string
     {
         return 'noscript';