+++ src/Tag/Title.php
@@ -4,16 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Html\Tag;
 
-use Yiisoft\Html\Tag\Base\NormalTag;
-use Yiisoft\Html\Tag\Base\TagContentTrait;
+use Yiisoft\Html\Tag\Base\ContentTag;
 
 /**
  * @link https://html.spec.whatwg.org/multipage/semantics.html#the-title-element
  */
-final class Title extends NormalTag
+final class Title extends ContentTag
 {
-    use TagContentTrait;
-
     protected function getName(): string
     {
         return 'title';